move parser to new file and add comments
